+++ src/main.ts
@@ -1,1 +1,10 @@
 export { default as DSButton } from './components/DSButton'
+export { default as DSInput } from './components/DSInput'
+export { default as DSIconButton } from './components/DSIconButton'
+export { default as DSMenu } from './components/DSMenu'
+export { default as DSMenuLink } from './components/DSMenuLink'
+export { default as DSList } from './components/DSInput'
+
+export * as DateUtils from './utils/DateUtils'
+export * as StringUtils from './utils/StringUtils'
+export * as useOutsideClick from './directives/useOutsideClick'