+++ src/types/menu_link.d.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from 'react'
+
+export interface MenuLinkProps {
+  href: string
+  noBorders?: boolean
+  children?: ReactNode
+  active?: 'on' | 'off'
+}