+++ src/utils/DateUtils.ts
@@ -1,0 +1,64 @@
+export const formatDateToStringDate = (date: Date): string => {
+  const month = String(date.getMonth() + 1).padStart(2, '0')
+  const year = date.getFullYear()
+  return `${month}/${year}`
+}
+
+export const formatDateToFullStringDate = (date: Date): string => {
+  const day = String(date.getDate()).padStart(2, '0')
+  const month = String(date.getMonth() + 1).padStart(2, '0')
+  const year = date.getFullYear()
+  return `${day}/${month}/${year}`
+}
+
+export const getNameOfDayOfTheWeek = (n: number): string => {
+  switch (n) {
+    case 0:
+      return 'Domingo'
+    case 1:
+      return 'Segunda-Feira'
+    case 2:
+      return 'Terça-Feira'
+    case 3:
+      return 'Quarta-Feira'
+    case 4:
+      return 'Quinta-Feira'
+    case 5:
+      return 'Sexta-Feira'
+    case 6:
+      return 'Sábado'
+    default:
+      return ''
+  }
+}
+
+export const getNameOfMonth = (n: number): string => {
+  switch (n) {
+    case 0:
+      return 'Janeiro'
+    case 1:
+      return 'Fevereiro'
+    case 2:
+      return 'Março'
+    case 3:
+      return 'Abril'
+    case 4:
+      return 'Maio'
+    case 5:
+      return 'Junho'
+    case 6:
+      return 'Julho'
+    case 7:
+      return 'Agosto'
+    case 8:
+      return 'Setembro'
+    case 9:
+      return 'Outubro'
+    case 10:
+      return 'Novembro'
+    case 11:
+      return 'Dezembro'
+    default:
+      return ''
+  }
+}