+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.0.3] - 2024-12-03
+
+### Adicionado
+
+- Novo componente criado - DSList.
+- Documentação do componente DSList criado no Storybook.
+- Novo componente criado - DSMenu.
+- Documentação do componente DSMenu criado no Storybook.
+- Novo componente criado - DSMenuLink.
+- Documentação do componente DSMenuLink criado no Storybook.
+- Criação de funções utils de Data.
+- Criação de funções utils de Strings.
+- Criação de diretiva para clique fora do componente.
+
+## [0.0.2] - 2024-11-28
+
+### Adicionado
+
+- Novo componente criado - DSInput.
+- Documentação do componente DSInput criado no Storybook.
+
 ## [0.0.1] - 2024-11-22
 
 ### Adicionado